Here are the results of merging the BigVars code. I hope that nothing has been lost or corrupted. I did find a few errors while checking the diffs.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11547 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/defs.h
@@ -3,11 +3,11 @@
  *      $Id: defs.h,v 1.27 2010-04-14 21:29:48 huangwei Exp $
  */
 /************************************************************************
-*									*
-*			     Copyright (C)  1993			*
-*	     University Corporation for Atmospheric Research		*
-*			     All Rights Reserved			*
-*									*
+*                                                                       *
+*                Copyright (C)  1993                                    *
+*        University Corporation for Atmospheric Research                *
+*                All Rights Reserved                                    *
+*                                                                       *
 ************************************************************************/
 /*
  *	File:		defs.h 
@@ -27,6 +27,7 @@ extern "C" {
 #define _NCdefs_h
 
 #include "ncarg/hlu/NgSizeT.h"
+#include    <stddef.h>
 
 #define NCL_MAX_DIMENSIONS 32
 #define NCL_MAX_FVARS 2048
@@ -82,10 +83,10 @@ typedef NrmQuark NclQuark;
 #define NclQuarkIsDef
 #endif
 
-typedef long NclValue;
+typedef size_t  NclValue;
 
 typedef struct _NclGenericVal {
-	int kind;
+	size_t  kind;
 	char *name;
 } NclGenericVal;
